Clean up Entity->spawnTo
