+++ src/pocketmine/entity/Arrow.php
@@ -42,10 +42,12 @@ class Arrow extends Projectile{
     protected $drag = 0.01;
     protected $damage = 2;
     protected $isCritical;
+
     public function __construct(FullChunk $chunk, Compound $nbt, Entity $shootingEntity = null, $critical = false){
         $this->isCritical = (bool) $critical;
         parent::__construct($chunk, $nbt, $shootingEntity);
     }
+
     public function onUpdate($currentTick){
         if($this->closed){
             return false;
@@ -69,22 +71,6 @@ class Arrow extends Projectile{
         }
         //$this->timings->stopTiming();
         return $hasUpdate;
-    }
-    public function spawnTo(Player $player){
-        if (!isset($this->hasSpawned[$player->getId()]) && isset($player->usedChunks[Level::chunkHash($this->chunk->getX(), $this->chunk->getZ())])) {
-            $this->hasSpawned[$player->getId()] = $player;
-            $pk = new AddEntityPacket();
-            $pk->type = static::NETWORK_ID;
-            $pk->eid = $this->getId();
-            $pk->x = $this->x;
-            $pk->y = $this->y;
-            $pk->z = $this->z;
-            $pk->speedX = $this->motionX;
-            $pk->speedY = $this->motionY;
-            $pk->speedZ = $this->motionZ;
-            //		$pk->metadata = $this->dataProperties;
-            $player->dataPacket($pk);
-        }
     }
 
     public function getBoundingBox() {
