+++ src/pocketmine/entity/PrimedTNT.php
@@ -142,22 +142,6 @@ class PrimedTNT extends Entity implements Explosive {
 		}
 	}
 
-	public function spawnTo(Player $player) {
-		$pk = new AddEntityPacket();
-		$pk->type = PrimedTNT::NETWORK_ID;
-		$pk->eid = $this->getId();
-		$pk->x = $this->x;
-		$pk->y = $this->y;
-		$pk->z = $this->z;
-		$pk->speedX = $this->motionX;
-		$pk->speedY = $this->motionY;
-		$pk->speedZ = $this->motionZ;
-		$pk->metadata = $this->dataProperties;
-		$player->dataPacket($pk);
-
-		parent::spawnTo($player);
-	}
-
 	public function setOwner($owner) {
 		$this->owner = $owner;
 	}
