+++ src/pocketmine/entity/Squid.php
@@ -79,25 +79,6 @@ class Squid extends WaterAnimal implements Ageable{
 	public function onUpdate($currentTick){
 	}
 
-
-	public function spawnTo(Player $player){
-		$pk = new AddEntityPacket();
-		$pk->eid = $this->getId();
-		$pk->type = Squid::NETWORK_ID;
-		$pk->x = $this->x;
-		$pk->y = $this->y;
-		$pk->z = $this->z;
-		$pk->speedX = $this->motionX;
-		$pk->speedY = $this->motionY;
-		$pk->speedZ = $this->motionZ;
-		$pk->yaw = $this->yaw;
-		$pk->pitch = $this->pitch;
-		$pk->metadata = $this->dataProperties;
-		$player->dataPacket($pk);
-
-		parent::spawnTo($player);
-	}
-
 	public function getDrops(){
 		return [
 			ItemItem::get(ItemItem::DYE, 0, mt_rand(1, 3))
